backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
kafka-python==2.0.2
pyiceberg==0.7.1
pyarrow==17.0.0
pandas==2.2.0
langchain==0.3.0
langchain-google-genai==2.0.0
faiss-cpu==1.8.0
sentence-transformers==3.0.0
textblob==0.18.0
websockets==12.0
pydantic==2.9.0
python-dotenv==1.0.0
